requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
wikipedia
duckduckgo-search
pydantic
python-dotenv